\name{BM_LL}
\alias{BM_LL}
\title{A function calculates the Log-Likelihood of the Mardia and Bindham desnity.}
\description{A function calculates the Log-Likelihood of the Mardia and Bindham desnity.}
\usage{BM_LL(param, u, v, mu.data)}
\arguments{
  \item{param}{A vector of parameters.}
  \item{u}{Directional cosine.}
  \item{v}{Directional cosine.}
  \item{mu.data}{A vector of means.}
}
\author{Nishan Mudalige}
\examples{
## To be added.
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory (show via RShowDoc("KEYWORDS")):
% \keyword{ ~kwd1 }
% \keyword{ ~kwd2 }
% Use only one keyword per line.
% For non-standard keywords, use \concept instead of \keyword:
% \concept{ ~cpt1 }
% \concept{ ~cpt2 }
% Use only one concept per line.
