\name{create_model}
\alias{create_model}
\title{A function to create a statistical model for length}
\description{A function to create a statistical model for length.}
\usage{create_model(data, include_names, u, v, trace = F)}
\arguments{
  \item{data}{The data to evaluate parameters.}
  \item{include_names}{Column names of data that should be included in the model.}
  \item{u}{Directional cosine.}
  \item{v}{Directional cosine.}
  \item{trace}{Whether to show the steps of the model selection process}
}
\author{Nishan Mudalige}
\examples{
## To be added.
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory (show via RShowDoc("KEYWORDS")):
% \keyword{ ~kwd1 }
% \keyword{ ~kwd2 }
% Use only one keyword per line.
% For non-standard keywords, use \concept instead of \keyword:
% \concept{ ~cpt1 }
% \concept{ ~cpt2 }
% Use only one concept per line.
