\name{fit_length}
\alias{fit_length}
\title{A function to fit the length of a tree branch}
\description{A function to fit the length of a tree branch}
\usage{fit_length(base, var, data, trace = F)}
\arguments{
  \item{base}{Formula of base model.}
  \item{var}{Variable to add to the model.}
  \item{data}{Data used for parameter estimation.}
  \item{trace}{Whether to show the steps of the model selection process.}
}
\author{Nishan Mudalige}
\examples{
## To be added.
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory (show via RShowDoc("KEYWORDS")):
% \keyword{ ~kwd1 }
% \keyword{ ~kwd2 }
% Use only one keyword per line.
% For non-standard keywords, use \concept instead of \keyword:
% \concept{ ~cpt1 }
% \concept{ ~cpt2 }
% Use only one concept per line.
