\name{getAIC_direction}
\alias{getAIC_direction}
\title{A function to obtain the AIC for the model for direction}
\description{A function to obtain the AIC for the model for direction}
\usage{getAIC_direction(x)}
\arguments{
  \item{x}{List of objects retrieved after nlminb optimization.}
}
\author{Nishan Mudalige}
\examples{
## To be added.
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory (show via RShowDoc("KEYWORDS")):
% \keyword{ ~kwd1 }
% \keyword{ ~kwd2 }
% Use only one keyword per line.
% For non-standard keywords, use \concept instead of \keyword:
% \concept{ ~cpt1 }
% \concept{ ~cpt2 }
% Use only one concept per line.
